Add Gulp 'clean' task
Added a Gulp 'clean' task to the beginning of the default Gulp task
list in order to automate clean-up of no longer generated Webpack files.
This revolves #20.

Also added JSDoc and coverage files clean-up to this 'clean' task's
implementation to enable developers to more easily be able to see just
the project source code.