Allow input with unparseable card indices
Enhance the application to allow entry of unknown card indices, treating
them as having undefined ordinals which can never contribute to hand
points, but still displaying the unknown values.  Useful for being
robust in the face of slightly bad user input (fixes #18), and possibly
in the case where a user is analyzing a hand in which not all cards are
known or fixed.